[fx] get rid of graph_module.root
instead submodules and weights are installed directly on the
graph_module by transferring the original modules. This makes it more
likely that scripting will succeed (since we no longer have submodules
that are not used in the trace). It also prevents layered transforms
from having to special case handling of the `root` module. GraphModules
can now be re-traced as part of the input to other transforms.

ghstack-source-id: 4d49f8d1ed8062a49931f9c30e7a422735013aa8
Pull Request resolved: https://github.com/pytorch/pytorch/pull/44092